+++ examples/spinning_cube/spinning_cube_app.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <cubey/app_config.h>
+
+namespace cubey::examples::spinning_cube {
+
+int run_spinning_cube(const RunConfig& config);
+
+} // namespace cubey::examples::spinning_cube